generating end to end tests from pict model

+++ GoogleTestAdapter/VsPackage.Tests/VsExperimentalInstance.cs
@@ -25,9 +25,6 @@ namespace GoogleTestAdapterUiTests
 
         public VsExperimentalInstance(Versions version, string suffix)
         {
-            if (string.IsNullOrWhiteSpace(suffix))
-                throw new ArgumentException("suffix may not be empty");
-
             Version = version;
             Suffix = suffix;
             VersionAndSuffix = $"{Version:d}.0{Suffix}";
@@ -44,6 +41,9 @@ namespace GoogleTestAdapterUiTests
 
         public void Clean()
         {
+            if (string.IsNullOrEmpty(Suffix))
+                throw new InvalidOperationException("We do not want to clean the non-experimental VS instance.");
+
             foreach (var dir in GetVsDirectories().Where(Directory.Exists))
                 Directory.Delete(dir, true);
             foreach (var key in GetVsHkcuKeys().Where(Registry.CurrentUser.HasSubKey))
@@ -75,15 +75,30 @@ namespace GoogleTestAdapterUiTests
             using (var settings = ExternalSettingsManager.CreateForApplication(GetExePath(), Suffix))
             {
                 var ems = new ExtensionManagerService(settings);
-                var vsix = ExtensionManagerService.CreateInstallableExtension(vsixPath);
+                IInstallableExtension vsix = ExtensionManagerService.CreateInstallableExtension(vsixPath);
+
+                if (ems.IsInstalled(vsix))
+                {
+                    IInstalledExtension installedVsix = ems.GetInstalledExtension(vsix.Header.Identifier);
+                    ems.Uninstall(installedVsix);
+                    if (ems.IsInstalled(vsix))
+                        throw new InvalidOperationException("Could not uninstall already installed GoogleTestAdapter.");
+                }
+
                 ems.Install(vsix, perMachine: false);
+                if (!ems.IsInstalled(vsix))
+                    throw new InvalidOperationException("Could not install GoogleTestAdapter.");
+
                 ems.Close();
             }
         }
 
         public Application Launch()
         {
-            return Application.Launch(new ProcessStartInfo(GetExePath(), $"/rootSuffix {Suffix}"));
+            ProcessStartInfo startInfo = string.IsNullOrEmpty(Suffix)
+                ? new ProcessStartInfo(GetExePath())
+                : new ProcessStartInfo(GetExePath(), $"/rootSuffix {Suffix}");
+            return Application.Launch(startInfo);
         }
 
         public static string GetVsTestConsolePath(Versions version)
